state debug info and add render dir
